# Improve MultiprocessIterator performance, functionality and stability, using Pool