… appveyor.yml
… tests/chainer_tests/iterators_tests/test_multiprocess_iterator.py