I rewrote the `MultiprocessIterator` using Python's `Pool` functionality. It makes codes simpler and faster, as well as making orphan/zombie process less likely.

Instead of previous "one prefetch invoked at one fetch" implementation, in my implementation background thread invokes prefetch freely, only if the maximum prefetch number allows. I think this strategy is better for utilizing idle CPU/IO time.

Resetting an iterator in the mid of iteration is also supported. Tests are changed accordingly.

The second commit changes per-batch-item shared memory allocation to the bulk allocation. This improved performance a little. 

I took a benchmark of `examples/mnist/train_mnist.py`, with `SerialIterator` changed to `MultiprocessIterator`.  It was done on EC2 g3.4xlarge instance.

```
time python train_mnist.py -g 0
baseline:          537.91s user 170.63s system 263% cpu 4:28.87 total
before bulk alloc: 234.32s user 27.87s  system 215% cpu 2:01.85 total
after bulk alloc:  227.92s user 26.92s  system 215% cpu 1:58.23 total
```