@okuta 2017-08-05
This PR fixes #2909 .