@niboshi 2017-08-05
@grafi-tt Could you add `TestMultiprocessIteratorConcurrency` test I wrote in #2919? It will check for reported deadlock (Issue #2909).