@niboshi 2017-08-05
Oh, the last commit of #2919 includes some debug code. So please use the 2nd-latest commit:

https://github.com/niboshi/chainer/blob/aea575ca60830fb23ce8fdbe0228c2d52ffb39d3/tests/chainer_tests/iterators_tests/test_multiprocess_iterator.py#L378