@okuta 2017-08-05
Could you merge [this PR](https://github.com/grafi-tt/chainer/pull/1/files) or fixing.