@grafi-tt 2017-08-06
@okuta Thank you for your fixes. Merged.