@grafi-tt 2017-08-06
@okuta The failure on AppVeyer is about maximum number of timeout at `therad.join()` on Windows environment. 
Possible fixes are:

1. use smaller number for timeout
2. set https://docs.python.org/3.6/library/threading.html#threading.TIMEOUT_MAX correctly for Python 3
3. do not join

I don't understand why join operation is required here.