@grafi-tt 2017-08-06
I found it's complicated than I thought. I'll think about this problem more, and wrote some fixes and tests (possibly).