@okuta 2017-08-06
If we do not use `join`, we cannot detect resource leak problem.
I want to provide stable framework.