@grafi-tt 2017-08-09
Interruption handling is tested now, whereas writing tests that work on Windows was tricky.
With Python 2.7 on Windows, wait for lock can't be interrupted even if long timeout is set. So I changed the implementation to periodically poll whether a termination request is sent. 1 second interval is compromise between performance and responsivity.