@grafi-tt 2017-08-10
@okuta @niboshi Now I think all issues are solved. Could you review changes again, and merge to the main tree if there is no problem?