@yuyu2172 2017-08-15
Is this checked to work with `MultiprocessPrallellUpdater`?
I tried your code to train a Convnet on ImageNet, and it always hangs when `MultiprocessParallelUpdater ` is used together.
The same code works fine with `StandardUpdater`.

I use Ubuntu 14.04 CUDA8.0, cuDNN 6.0, and basically my script is same as `examples/imagenet/train_imagenet_data_parallel.py`.