@grafi-tt 2017-08-25
@okuta @yuyu2172 I confirmed the issue using the imagenet example, and it is fixed by the last commit. Could you test your code again with this commit?

Fundamental cause is that `multiprocessing.Process` and `multiprocessing.Pool` are *not* thread safe if process starting method is "fork", which is default on unix. In this case, background threads will remain on forked processes.