@grafi-tt 2017-09-05
@okuta Fixed.

Though there is [an error on Travis CI Python 3.5 Linux build](https://travis-ci.org/chainer/chainer/jobs/271901409), I suppose it is not a problem. This error is a numerical mismatch seemingly that is slightly beyond the threshold. It is unrelated to `MultiprocessIterator`.