@grafi-tt 2017-09-11
@okuta Thank you very much for you detailed review. I checked error messages caused by interruption on AppVeyor (Windows). The result was as expected so I disabled this cluttering messages again.

If there are no other issues, I think you can merge it.